+++ packages/core/src/components/Accordion/Accordion/__stories__/Accordion.stories.tsx
@@ -61,7 +61,14 @@ export default {
 export const Overview = {
   render: accordionTemplate.bind({}),
   name: "Overview",
-  args: {}
+  args: {},
+  parameters: {
+    docs: {
+      liveEdit: {
+        isEnabled: false
+      }
+    }
+  }
 };
 
 export const MultiActive = {