+++ packages/core/src/components/Tabs/Tab/__stories__/Tab.stories.tsx
@@ -15,19 +15,26 @@ export const Overview = {
   args: {
     children: "Tab"
   },
-  name: "Overview"
+  name: "Overview",
+  parameters: {
+    docs: {
+      liveEdit: {
+        isEnabled: false
+      }
+    }
+  }
 };
 
 export const States = {
   render: () => (
     <>
       <Tab>Normal</Tab>
-      <Tab disabled>Disabled</Tab>
+      <Tab disabled tooltipProps={{ content: "This tab is disabled" }}>
+        Disabled
+      </Tab>
       <Tab active>Active</Tab>
     </>
-  ),
-
-  name: "States"
+  )
 };
 
 export const Icons = {
@@ -39,6 +46,11 @@ export const Icons = {
       </Tab>
     </>
   ),
-
-  name: "Icons"
+  parameters: {
+    docs: {
+      liveEdit: {
+        scope: { Email }
+      }
+    }
+  }
 };