+++ packages/testkit/components/Menu.ts
@@ -23,33 +23,19 @@ export class Menu extends BaseElement {
   }
 
   /**
-   * Initialize list items if they are not already initialized.
-   * @returns {Promise<void>}
+   * Get all menu items.
+   * @returns {Promise<MenuItem[]>} An array of MenuItem objects.
    */
-  async initializeItemsIfNeeded(): Promise<void> {
-    await test.step(`Initialize ${this.elementReportName} if needed`, async () => {
-      if (!this.itemsInitialized) {
-        await this.initializeItems();
-        this.itemsInitialized = true;
-      }
-    });
-  }
-
-  /**
-   * Initialize the list items by locating all list elements.
-   * @returns {Promise<void>}
-   */
-  async initializeItems(): Promise<void> {
-    await test.step(`Initialize ${this.elementReportName}`, async () => {
-      await this.waitForElementsGroup(this.locator.locator("[role='menuitem']"), this.elementReportName);
-      const listElements = await this.locator.locator("[role='menuitem']").all();
-      this.items = await Promise.all(
-        listElements.map(async locator => {
-          const itemName = await locator.innerText();
-          return new MenuItem(this.page, locator, `Menu Item: ${itemName}`);
-        })
+  async getAllMenuItems(): Promise<MenuItem[]> {
+    let menuItems: MenuItem[] = [];
+    await test.step(`Get all menu items in ${this.elementReportName}`, async () => {
+      const menuItemsLocators = await this.locator.locator("[role='menuitem']").all();
+      const menuItemsPromises = menuItemsLocators.map(
+        async locator => new MenuItem(this.page, locator, await locator.innerText())
       );
+      menuItems = await Promise.all(menuItemsPromises);
     });
+    return menuItems;
   }
 
   /**
@@ -57,8 +43,16 @@ export class Menu extends BaseElement {
    * @param {string} itemName - The name of the item to retrieve.
    * @returns {ListItem | undefined} The list item with the specified name or undefined if not found.
    */
-  getItemByName(itemName: string): MenuItem | undefined {
-    return this.items.find(item => item.elementReportName.includes(itemName));
+  async getItemByName(itemName: string): Promise<MenuItem | undefined> {
+    let menuItem: MenuItem | undefined;
+    await test.step(`Get menu item by name ${itemName} in ${this.elementReportName}`, async () => {
+      menuItem = new MenuItem(
+        this.page,
+        this.locator.locator("[role='menuitem']").filter({ hasText: itemName }),
+        `Menu Item: ${itemName}`
+      );
+    });
+    return menuItem;
   }
 
   /**
@@ -67,13 +61,11 @@ export class Menu extends BaseElement {
    * @returns {Promise<void>}
    */
   async selectItem(listItem: string): Promise<void> {
-    await this.initializeItemsIfNeeded();
-    let menuItem = this.getItemByName(listItem);
-    if (!menuItem) {
-      // If the item is not in the initialized list, create it dynamically
-      menuItem = new MenuItem(this.page, this.locator.getByText(`${listItem}`), `Menu Item: ${listItem}`);
-    }
-    await menuItem.scrollIntoView();
+    const menuItem = new MenuItem(
+      this.page,
+      this.locator.locator("[role='menuitem']").filter({ hasText: listItem }),
+      `Menu Item: ${listItem}`
+    );
     await menuItem.click();
   }
 }